requirements.txt
djangorestframework
markdown
django-filter
psycopg2
rest_framework_simplejwt
corsheaders
pillow
pytz